refactor auth spec separating create user logic
